mailing/apps/users/api/serializers.py
from rest_framework import serializers

from mailing.apps.users.models import User

class UserSerializer(serializers.ModelSerializer):
    """Serializer for user registration."""

    class Meta:
        model = User
        fields = ('id', 'name', 'email', 'password')
        extra_kwargs = {'password': {'write_only': True}}
