requirements.txt
Django==3.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1